+++ compiler/__init__.py
@@ -1,14 +1,15 @@
 from platform import system
 from subprocess import run
+from pprint import pprint
 from typing import Union
 from pathlib import Path
 from shutil import which
 
-from antlr4 import FileStream, CommonTokenStream
+from antlr4 import InputStream, CommonTokenStream
 
 from compiler.info import (
-    get_info, op_name_to_symbol_map, DEFAULT_HEADER, is_function, is_type, is_method, is_property,
-    is_static
+    get_info, op_name_to_symbol_map, DEFAULT_HEADER, is_method, is_property,
+    is_static, has_overload
 )
 from compiler.constants import (
     Code, EnvItem, Position, BOLD, RESET, RED, find_function, base_type, init_env
@@ -28,12 +29,21 @@ class SplikitCompiler(SplikitVisitor):
         self.includes = set()
 
         self.info = get_info()
+    
+    @property
+    def error_checker(self) -> Union[BasicErrorChecker, None]:
+        return self._error_checker if hasattr(self, '_error_checker') else None
+    
+    def compile_string(self, src: str, compiler: Union['SplikitCompiler', None] = None) -> str:
+        sub_env = self.env.copy()
+        init_env(sub_env, self.info)
+        
+        if compiler is None:
+            compiler = SplikitCompiler(sub_env)
 
-    def compile(self, file: Path) -> Path:
-        self.src = file.read_text('utf-8')
-        fstream = FileStream(file.as_posix())
+        istream = InputStream(src)
 
-        lexer = SplikitLexer(fstream)
+        lexer = SplikitLexer(istream)
         tokens = CommonTokenStream(lexer)
 
         parser = SplikitParser(tokens)
@@ -41,13 +51,21 @@ class SplikitCompiler(SplikitVisitor):
         parser.addErrorListener(SplikitErrorListener(self.src))
 
         tree = parser.parse()
-        
+
+        self._error_checker = BasicErrorChecker(sub_env.copy(), self.src)
+        self._error_checker.info = compiler.info
+        self._error_checker.visit(tree)
+        # pprint(self._error_checker.info)
+
+        return compiler.visit(tree)
+
+    def compile(self, file: Path, extension: str = '.cpp') -> Path:
+        self.src = file.read_text('utf-8')
         init_env(self.env, self.info)
 
-        BasicErrorChecker(self.env.copy(), self.src).visit(tree)
-        cpp_code = self.visit(tree)
+        cpp_code = self.compile_string(self.src, self)
         
-        cpp_file = file.with_suffix('.cpp').absolute()
+        cpp_file = file.with_suffix(extension).absolute()
         includes = f'#include "{DEFAULT_HEADER.absolute().as_posix()}"\n'
         includes += ''.join(f'#include "{inc.absolute().as_posix()}"\n' for inc in self.includes)
         cpp_file.write_text(f'{includes}\n\n{cpp_code}', 'utf-8')
@@ -71,12 +89,60 @@ class SplikitCompiler(SplikitVisitor):
             expr.type = 'bool'
 
         return expr
-
-    def body_to_str(self, ctx: SplikitParser.BodyContext) -> str:
-        return '\n'.join(map(lambda x: x.text + ';', self.visit(ctx)))
     
     def get_text(self, iterable: list) -> str:
         return ', '.join(map(lambda x: x.text, iterable))
+    
+    def body_to_str(self, ctx: SplikitParser.BodyContext) -> str:
+        return '\n'.join(map(lambda x: x.text + ';', self.visit(ctx)))
+    
+    def interpolate_string(self, s: str, position: Position) -> str:
+        if s.startswith('$'):
+            i = 0
+            output_string = ''
+            string_type = s[1]
+            error_message = f'Use of {string_type} in outer string, can\'t use in inner string'
+            s = s[2:-1]
+            while i < len(s):
+                if i + 1 > len(s):
+                    continue
+                
+                char = s[i]
+                
+                if char == '{':
+                    concatenation = ''
+                    for c in s[i+1:-1]:
+                        if c == '}':
+                            break
+                        elif c == string_type:
+                            position.error_here(error_message, self.src)
+
+                        concatenation += c
+
+                        i += 1
+
+                    code = self.compile_string(concatenation, self)
+                    output_string += f'" + repr({code}) + "'
+                    i += 1
+                else:
+                    output_string += char
+
+                i += 1
+            
+            return f'"{output_string}"'
+        
+        return f'"{s[1:-1]}"'
+
+    def get_body(self, ctx: SplikitParser.BodyContext, params: list[Code]) -> str:
+        temp_env = self.env.copy()
+        for param in params:
+            splitted = param.text.split()
+            self.env[splitted[1]] = EnvItem(splitted[1], splitted[0], False)
+        
+        body = self.body_to_str(ctx)
+        
+        self.env = temp_env
+        return body
 
 
     def visitParse(self, ctx: SplikitParser.ParseContext) -> str:
@@ -114,6 +180,22 @@ class SplikitCompiler(SplikitVisitor):
 
     def visitBody(self, ctx: SplikitParser.BodyContext) -> list[Code]:
         return [self.visit(body) for body in ctx.bodyStmts()]
+
+    def visitForeachStmt(self, ctx: SplikitParser.ForeachStmtContext) -> Code:
+        var = ctx.ID().getText()
+        expr = self.visit(ctx.expr())
+        
+        temp_env = self.env.copy()
+        self.env[var] = EnvItem(var, expr.type, False)
+        
+        body = self.body_to_str(ctx.body())
+        self.env = temp_env
+        
+        return Code(
+            f'for (auto {var} : {expr.text}) {{\n{body}\n}}',
+            'nil',
+            Position(ctx.start.line, ctx.start.column)
+        )
 
     def visitIfStmt(self, ctx: SplikitParser.IfStmtContext) -> Code:
         condition = self.condition(ctx.expr())
@@ -180,24 +262,12 @@ class SplikitCompiler(SplikitVisitor):
                 header = LIBS[path]
                 self.includes.add(header)
                 lib_info = get_info(header)
-                
-                new_info = {}
-                for k, v in lib_info.items():
-                    if is_type(k, v) or is_function(v):
-                        continue
-
-                    new_info[k] = v
-
-                for _, v in lib_info.items():
-                    if 'function' in v[0].split():
-                        self.env[v[2]] = EnvItem(v[2], v[1], False)
-
-                init_env(self.env, new_info)
-                self.info |= new_info
+                if self.error_checker is not None:
+                    self.error_checker.add_info(self, lib_info)
             else:
                 position.error_here(f'Unknown library \'{path}\'', self.src)
 
-        return Code('', 'nil', position)
+        return Code(f'// {", ".join(paths)} used', 'nil', position)
 
     def visitVarAssign(self, ctx: SplikitParser.VarAssignContext) -> Code:
         name = ctx.ID().getText()
@@ -237,19 +307,12 @@ class SplikitCompiler(SplikitVisitor):
             param_dict[splitted[1]] = {'type': {splitted[0]}}
 
         @std_func(param_dict)
-        def f(_, _env: dict, _call_position: Position) -> None:
+        def f(*_) -> None:
             pass
 
         self.env[name] = EnvItem(name, typ, False, f)
 
-        temp_env = self.env.copy()
-        for param in params:
-            splitted = param.text.split()
-            self.env[splitted[1]] = EnvItem(splitted[1], splitted[0], False)
-
-        body = self.body_to_str(ctx.body())
-
-        self.env = temp_env
+        body = self.get_body(ctx.body(), params)
         inlined = 'inline ' if ctx.INLINE() else ''
         return Code(
             f"""{inlined}{typ} {name}({self.get_text(params)}) {{
@@ -262,29 +325,54 @@ class SplikitCompiler(SplikitVisitor):
 
     def visitAtom(self, ctx: SplikitParser.AtomContext) -> Code:
         position = Position(ctx.start.line, ctx.start.column)
-        if ctx.INT():
+        if ctx.INT() is not None:
             return Code(ctx.getText(), 'int', position)
-        elif ctx.FLOAT():
+        elif ctx.FLOAT() is not None:
             return Code(ctx.getText() + 'f', 'float', position)
-        elif ctx.STRING():
-            return Code(ctx.getText(), 'string', position)
-        elif ctx.BOOL():
+        elif ctx.STRING() is not None:
+            return Code(self.interpolate_string(ctx.getText(), position), 'string', position)
+        elif ctx.BOOL() is not None:
             return Code(ctx.getText(), 'bool', position)
-        elif ctx.NIL():
+        elif ctx.NIL() is not None:
             return Code(ctx.getText(), 'nil', position)
-        elif ctx.ID():
+        elif ctx.ID() is not None:
             return Code(ctx.getText(), self.env[ctx.ID().getText()].type, position)
-        elif ctx.expr():
+        elif ctx.expr() is not None:
             expr = self.visit(ctx.expr())
             expr.text = f'({expr.text})'
             return expr
-        elif ctx.LBRACE():
+        elif ctx.LBRACE() is not None:
             typ = self.visit(ctx.type_())
             args = self.visit(ctx.args()) if ctx.args() else []
             return Code(
                 f'new_array<{typ}>({{{self.get_text(args)}}})',
                 f'array<{typ}>', position
             )
+        elif ctx.REGEX() is not None:
+            return Code(
+                f'new_regex("{ctx.getText()[1:-1].replace("\\", "\\\\")}")',
+                'regex', position
+            )
+    
+    def _inline_call(
+        self, info_name: str, args: list[Code], func_name: Union[str, None] = None
+    ) -> str:
+        if func_name is None:
+            func_name = info_name
+
+        if info_name in self.info:
+            f = self.info[info_name]
+            code = ''
+            for comment in f.comments:
+                if comment.startswith('single-return('):
+                    comment = comment[14:-1]
+                    code = comment
+                    for i, arg in enumerate(args, start=1):
+                        code = code.replace(f'@{i}@', arg.text)
+
+                    return code
+
+        return f'{func_name}({self.get_text(args)})'
 
     def visitCall(self, ctx: SplikitParser.CallContext) -> Code:
         callee = self.env.get(ctx.ID().getText())
@@ -305,9 +393,9 @@ class SplikitCompiler(SplikitVisitor):
         return list(op_name_to_symbol_map.keys())[list(op_name_to_symbol_map.values()).index(op)]
 
     def visitExpr(self, ctx: SplikitParser.ExprContext) -> Code:
-        if ctx.atom():
+        if ctx.atom() is not None:
             return self.visit(ctx.atom())
-        elif ctx.call():
+        elif ctx.call() is not None:
             return self.visit(ctx.call())
         elif ctx.op:
             left = self.visit(ctx.expr(0))
@@ -318,31 +406,31 @@ class SplikitCompiler(SplikitVisitor):
 
             position = Position(ctx.op.line, ctx.op.column)
             
-            # manually add int divided by int because C++ already has a div(int, int) and it cannot
-            # be overriden
+            # manually add int / int, C++ already has a div(int, int) and it cannot be overriden
             if op == '/' and left_type == 'int' and right_type == 'int':
-                return Code(f'div({left.text}, {right.text}).quot', 'int', position)
+                return Code(f'{left.text} / {right.text}', 'int', position)
             elif left_type == 'any' and right_type == 'any':
                 typ = left.type
             else:
-                info_item = self.info.get((op, (left_type, right_type)))
-                typ = info_item[1]
+                info_item = find_function(self.info, op)
+                typ = info_item.return_type
 
             name = self._op_symbol_to_name(op)
-            return Code(f'{name}({left.text}, {right.text})', typ, position)
+            return Code(self._inline_call(op, [left, right], name), typ, position)
         elif ctx.NOT():
             left = self.visit(ctx.expr(0))
             left_type = base_type(left.type)
-            info_item = self.info.get(('!', (left_type,)))
+            info_item = find_function(self.info, '!')
             position = Position(ctx.start.line, ctx.start.column)
-            if info_item is None:
+            if info_item is None or not has_overload(info_item, [left_type]):
+                call = self._inline_call('to_bool', [left])
                 return Code(
-                    f'!to_bool({left.text})',
+                    f'!{call}',
                     'bool',
                     position
                 )
 
-            return Code(f'not_({left.text})', info_item[1], position)
+            return Code(self._inline_call('!', [left], 'not_'), info_item.return_type, position)
         elif ctx.DOT():
             left = self.visit(ctx.expr(0))
             ltype = left.type
@@ -355,33 +443,30 @@ class SplikitCompiler(SplikitVisitor):
             if info_item is None:
                 position.error_here(f'\'{ltype}\' has no attribute \'{attr}\'', self.src)
 
-            # if info_item[2].split(', ')[0].split()[0].startswith('array<') and ltype == 'array':
-
             args = []
             if not is_static(info_item):
                 args.append(left)
 
             if is_property(info_item):
+                call = self._inline_call(f'{ltype}_{attr}', args)
                 return Code(
-                    f'{ltype}_{attr}({self.get_text(args)})',
-                    info_item[1],
+                    call,
+                    info_item.return_type,
                     position
                 )
             elif is_method(info_item):
                 args.extend(self.visit(ctx.args()) if ctx.args() else [])
                 params_dict = {}
-                for param in info_item[3].split(', '):
-                    if param == '':
-                        continue
-
-                    splitted = param.split()
-                    params_dict[splitted[1]] = {'type': {splitted[0]}}
+                for name, typ in zip(info_item.overloads[0][1], info_item.overloads[0][0]):
+                    params_dict[name] = {'type': {typ}}
 
                 verify_params(args, params_dict, self, position)
+                call = self._inline_call(f'{ltype}_{attr}', args)
                 return Code(
-                    f'{ltype}_{attr}({self.get_text(args)})',
-                    info_item[1],
+                    call,
+                    info_item.return_type,
                     position
                 )
-            else:
-                position.error_here(f'\'{attr}\' is an invalid attribute', self.src)
+        elif ctx.SUB() is not None:
+            expr = self.visit(ctx.expr(0))
+            return Code(ctx.getText(), expr.type, Position(ctx.start.line, ctx.start.column))