+++ compiler/std/fstream.hpp
@@ -18,26 +18,26 @@ inline string type(File& _) { return "File"; }
 
 inline string repr(File& f) { return "File(path='" + f.path + "')"; }
 
-// function
-File open_file(string path) {
-    return {std::fstream(path), std::filesystem::path(path), path};
-}
+// function, single-return({std::fstream(@1@),std::filesystem::path(@1@),@1@})
+File open_file(string path) { return {std::fstream(path), std::filesystem::path(path), path}; }
 
 
-// property
+// property, single-return(@1@.path)
 inline string File_path(File& f) { return f.path; }
 
-// property
+// property, single-return(std::filesystem::is_regular_file(@1@.p))
 inline bool File_is_file(File& f) { return std::filesystem::is_regular_file(f.p); }
 
-// property
+// property, single-return(std::filesystem::is_directory(@1@.p))
 inline bool File_is_dir(File& f) { return std::filesystem::is_directory(f.p); }
 
-// property
+// property, single-return(std::filesystem::exists(@1@.p))
 inline bool File_exists(File& f) { return std::filesystem::exists(f.p); }
 
+// single-return(File_exists(@1@))
 inline bool to_bool(File& f) { return File_exists(f); }
 
+// single-return(!File_exists(@1@))
 inline bool not_(File& f) { return !to_bool(f); }
 
 // property